check: add checked pointer autodereference function
To be used in all the places that do autoderef and have to check that
they're not doing it on a nullable pointer. The only visible change in
this patch is a slightly more generic error message in all cases, but
the extra verbosity did not provide much value since the offending pointer is
clear from the location.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>