check: remove switching on pointers
It's not really implementable in a general way without risking duplicate
cases that cannot be detected at compile time. A limited form in which
only switching on pointers inside a single object is allowed might be
workable if someone really wanted it, but it is of questionable utility
because at that point one can just use offsets.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>