gen: generalize non-C-ABI-compatible type support
This emits an opaque type unconditionally if we end up with some kind of
type which is not straightforwardly representable in the C ABI (or, to
be blunt, the qbe grammar, which is not the same thing but alas).

It is possible that the assertion previously shown here would fire on
normative conditions when mixing non-C-ABI types with an aggregate type,
for instance an array of @packed structs.

This fixes a regression introduced by
a26eb9ad9672cdc950d4bb9ef9e39f34a50dd817,
"type_store: make padding calculation explicit where required". Or
rather it fixes a bug that this commit drew my attention to.

Signed-of-by: Drew DeVault <sir@cmpwn.com>