handle allocation failure by returning nomem
previously, if the result type of an allocation-expression was not a
nullable pointer type, the builtin would abort. this is now handled
by instead returning nomem:

        let x: (*int | nomem) = alloc(1);

the same goes for append and insert:

        let x: (void | nomem) = append(...);
        let y: (void | nomem) = insert(...);

nomem is a zero-sized type (like void). it's error flag is
always set.

this commit introduces the following breaking changes:

        - nomem cannot be used as a name anymore
        - alloc, insert and append have different return values now
        - harec binary interface has changed.

Co-authored-by: Bor Grošelj Simić <bgs@turminal.net>
Signed-off-by: Lorenz (xha) <me@xha.li>
Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>