gen: handle a switch corner case
Before 56f9fa50, switch expressions had result type never if all of
their branches had result never. This was against the spec in cases
where the switch was provided a non-never type hint.

Now the result type matches what the spec dictates, but this created a
problem in gen, where, based on the non-never type hint, a result is
expected from at least one of the branches. To make QBE happy, a dummy
initialization of the switch result is added. It has no practical
impact, since QBE is able to optimize it away in all cases.

Reported-by: Ember Sawady <ecs@d2evs.net>