parse: improve empty block error message
"syntax error: cannot have empty block" used to sometimes be reported
when it shouldn't have been. It was reported whenever a closing brace
was used where an expression was expected. So something like this would
report it:

	if (foo) {
		func(thing1,
			thing2,
			thing3,
	};

This is fixed by only reporting this error when an opening brace is
immediately followed by a closing brace.

Additionally, a similar error message is now also used for empty cases.

Finally, the error location now points to the opening brace, rather than
the closing brace. This is consistent with the empty case error, which
points to the `case` token preceding the empty case, rather than to the
following `case` token or closing brace.