Implement discarding bindings and assignment.
	let _ = expr;
	_ = expr;
	for (let _ .. expr) void;
	for (let _ &.. expr) void;
	for (let _ => expr) void;

Both binding and assignment forms are allowed. The assignment form is an
expression and can replace void casts (eg if (cond) _ = can_fail()). The
binding form is included for consistency with tuple assignments and
for-each binding lists.

Also, remove the ability to cast any expression to void. This C idiom is
not allowed by the spec and is no longer necessary with discarding
bindings.

	can_fail(): void becomes _ = can_fail();

Void aliases and void tagged union members may still be cast as
specified.

Breaking-Change: language
Implements: https://todo.sr.ht/~sircmpwn/hare/461
Signed-off-by: Joe Finney <me@spxtr.net>