gen: extend bools for loops and asserts.
The code

	fn foo() bool;

	export fn main() void = {
		assert(foo());
	};

Along with "for (foo()) ..." was leading to qbe of the form

	%.8 =w call $foo()
	jnz %.8, @passed.7, @failed.6

However, foo() returns bool, not word. The upper bits of the returned
bool may be set, it is only the lowest byte that matters. In harec,
"if" expressions catch this with a call to an extend function that
inserts an extub instruction between the call and the jnz. In this patch
I also insert extend for for and for assert.

I thought that the correct fix was to use =ub instead of =w, but if I'm
reading the QBE docs right, that does absolutely nothing, which I find
strange.

This kind of bug worries me, because it seems there's a lot of ways
subword return values from foreign functions can creep in and make code
subtly broken.

Reported-By: elb <elb@disroot.org>
Signed-off-by: Joe Finney <me@spxtr.net>